+++ src/ArduinoJson/Array/ArrayRef.hpp
@@ -62,6 +62,7 @@ class ArrayRefBase {
 };
 
 class ArrayConstRef : public ArrayRefBase<const CollectionData>,
+                      public VariantOperators<ArrayConstRef>,
                       public Visitable {
   friend class ArrayRef;
   typedef ArrayRefBase<const CollectionData> base_type;
@@ -116,6 +117,7 @@ class ArrayConstRef : public ArrayRefBase<const CollectionData>,
 
 class ArrayRef : public ArrayRefBase<CollectionData>,
                  public ArrayShortcuts<ArrayRef>,
+                 public VariantOperators<ArrayRef>,
                  public Visitable {
   typedef ArrayRefBase<CollectionData> base_type;